[r5sim] Add vdisk and improve machine and simple_core
Clean up the half assed tracing/debugging log message support; it's
still not really scalable, but it's definitely not as bad as it was
before.

Also expand a bit on the vuart and the "driver" for it in the BROM.
Add a new VDISK device that acts a bit like a super simple block
device. With this we can now load a "bootloader".

Next steps:

  1. Implement the M extension (integer multiply/divide).
  2. Port u-boot (this seems to need the M extension).

Signed-off-by: Alex Waterman <alexw@nvidia.com>