Lift core->exec() to core.c
Move the core->exec() function to a global function in core.c
called r5sim_core_exec(). This function will now be responsible
for calling a core->exec_one() defined by a core implementation.

By moving the core_exec functionality to the generic core code
it will be easier to implement interrupts at a machine level.
The new r5sim_core_exec() function will be able to handle the
necessary interrupt pending code and dispatch an interrupt to
the core implementation - if it wishes to receive it.

Signed-off-by: Alex Waterman <alexw@nvidia.com>