[r5sim] Add SYSTEM instructions and basic CSRs
Add proper decoding for SYSTEM instructions along with two of the
basic CSRs required by rv32i: cycle[h] and instret[h].

This _appears_ to also add time[h] but this actually does nothing
at the moment. This needs to be tied to a read based event.

Also add some super basic tests for the rdcycle and rdinstret
psuedo instructions.

Signed-off-by: Alex Waterman <imnotlistening@gmail.com>