r5sim: conftest: Add trivial MIPS calculation
Now that we have the M extension we can do a quick IPS calculation
to work out just how slo - I mean fast - this simulator is. On my
machine with a AMD Ryzen 9 3950X I a little under 10 MIPS. There may
be a spool up time for the CPU executing the simulator to ramp up
clocks and stuff, but probably it's not a huge effect.

I can add some more benchmarks in time. But for now this was an
interesting experiment.

Signed-off-by: Alex Waterman <imnotlistening@gmail.com>