r5sim: core: Add the beginning of MMODE trap handling
Add preliminary support for MMode trap handling. This covers just
traps, not interrupts - that will come later. To do this a fair
amount of code had to be added.

Firstly move main DRAM from 1GB to 512MB; this is required because
it seems as though it's impossible to install a trap vector at a
address above 1GB due to MTVEC.BASE is only 30 bits.

Also modify the .exec_one() function to return an exception value
corresponding to a synchronous trap. This works for synchronous
traps since the instruction that traps is not executed. There's
a special case here for the MRET instruction; this is handled by
the actual core code.

Also note that only the MMODE trap handling - and a subset of
that - is actually implemented.

Lastly add some testing code to verify that traps are called and
returned from in the conftest. This just tests an illegal
instruction and a ECALL instruction.

Signed-off-by: Alex Waterman <imnotlistening@gmail.com>