r5sim: brom: Use volatile for readl(), writel()
For some reason, today, the compiler decided to start reordering
the readl() used by getc(); this caused a crash in the brom for
some reason. I'm not sure why, to be honest, but once the correct
ordering was enforced for the brom getc() call, things started
working again.

I found this bug since I noticed that the bootloader load was
executing before the getc() that's normally be used to let a user
choose when the brom actually executes.

Signed-off-by: Alex Waterman <imnotlistening@gmail.com>