r5sim: Support interrupts (sort of)
Begin support for interrupts. This patch modifies the main execution
loop such that it can be called recursively. Interrupts and traps,
when detected, break the inner while loop of the r5sim_core_exec()
function. The outer loop then calls r5sim_core_exec() again by way
of the trap and interrup functions.

An MRET instuction triggers the r5sim_core_exec() loop to immediately
break, which returns to the core trap handler and unwinds the trap.
Finally the original while loop starts running again.

There's also a fair amount of other stuff added:

  o Extra conf tests,
  o Updated CSR HW defs so that conftest and the r5sim core
    can share the defs,
  o A SW interrupt trigger in vsys,
  o Interrupt handling support in conftest

Some details on the interrupt handling internals from the device
perspective. Each core has a queue of incoming interrupts; that way
each signaled interrupt is guaranteed to interrupt the processor
exactly once - this probably doesn't really reflect reality, if I
had to guess.

When a core detects an incoming interrupt by seeing that the
interrupt queue is not empty, it dequeues the interrupt and, assuming
any interrupts are unmasked, triggers an interrupt. Oops! This is a
"HW" bug in that we'll trigger if _any_ interrupt is unmasked, not
just the specific dequeued interrupt.

Signed-off-by: Alex Waterman <imnotlistening@gmail.com>