r5sim: Add m-mode timer interrupts
Add machine mode timer interrupts triggerable from the vsys IO
device.

Technically this timer does not comport with the timer specified
in the RISC-V ISA; that can be implemented later. To me, at least,
this timer architecture is a bit easier to work with in SW.

Also add a WFI instruction in the simple core and a WFI
implementation in the core_intr code; all this does is wait until
the intr-queue is non-empty. Once the queue has an item in it the
WFI returns, the WFI instruction finishes executing, and the core
will handle the interrupt.

As a result the MEPC is corectly set pointing to after the WFI
instruction (would be bad if we jusrt immediately WFI'ed again!).
There's some other details about the WFI instruction when
interrupts are masked that I haven't sorted through yet. In
particular:

  > implementations should avoid resuming the hart if the
  > interrupt is pending but not individually enabled.

This, currently is not being checked. It'd require peeking at the
pending interrupt which has yet to be implemented.

Signed-off-by: Alex Waterman <imnotlistening@gmail.com>