r5sim: Update build approach for R5 code
For the R5 binaries, split the compile and link phases completely.
Now there's an explicit LD call as opposed to relying on GCC to do
it for us. This is a first step for building the R5 binaries in a
multi step process:

  1. Compile all the object files.
  2. Phase 1 link: build a temporary object that is the sum of
     all of the objects.
  3. Generate a symbol table containing all the symbols in our
     phase 1 link.
  4. Parse to C and then compile the symbol table into an object
     file. Place this in a special section.
  5. Link the resulting object file and the phase 1 link.

(5) is still TBD; not sure how exactly to do this. But if I can
get this to work, then we can implement back traces with symbols
at run time!

Signed-off-by: Alex Waterman <imnotlistening@gmail.com>