conftest: Add support for symbolic backtrace prints
With a little help from the Linux kernel RISC-V code and multiphase
kallsyms linking, implement some code to look up symbols that get
built and stored during linking. Check the kallsyms man page for
some more details on how the link processing works but the TL;DR is
that it takes 4 link runs.

Also add a test to make sure the back tracing works. It's a pretty
simple test that really serves mostly as just a quick smoke test.

Lastly modify the general makefile build flow a bit more to properly
support this.

Signed-off-by: Alex Waterman <imnotlistening@gmail.com>