r5sim: Add supervisor mode
Completely redesign the interrupt handling mechanism and with that
support the beginnings of supervisor mode. At this point the r5sim
now supports M-mode and S-mode along with what I think is correct
trap handling with respect to delegation and priv levels.

Also add a few tests to see if at least basic m-mode and s-mode
delegation and calling works.

Signed-off-by: Alex Waterman <imnotlistening@gmail.com>