r5sim: More build changes; now it works in parallel
Make a recurse.mk makefile to first include the expected Makefile
and then the rules. This makes it so that we can make directory
targets based on the subdir names that show up in $(OBJS).

By having targets for each of the subdirs make can execute
things in parallel where they have no dependencies. Thus the build
time is now a lot faster (not that it was that bad to begin with
but could be useful in the future).

Signed-off-by: Alex Waterman <imnotlistening@gmail.com>