r5sim: Add break command to debugger
Add a break command - and break points! - to the debugger and the
simulator. Machines now have a set of 4 HW break points that the
core implementation checks against every instruction. The number of
breaks is limited to keep try anbd keep the amount of overhead to
a reasonable minimum.

Also delete all the messy and stupid pthreading stuff for the
debugger. Now the debugger is far simpler and the only thing the
CTRL-z singal handler does is set a variable in the machine.

Signed-off-by: Alex Waterman <imnotlistening@gmail.com>