r5sim: Return better errors from mach->mem*()
mach->mem*() functions have no idea their context so cannot actually
return descriptive errors. For example how does the memload know
whether it's from an instruction or a regular data load?

As a result we have to return generic errors at the mach level and then
assign the proper trap exception return in the simple core. The
previous patch was broken for instruction load errors, although it
probably wouldn't have manifested since, in theory, a PC load should
never actually be misaligned.

In any event, this is a bit closer to the R5 spec, I think.

Signed-off-by: Alex Waterman <imnotlistening@gmail.com>