conftest: Add more exception tests
Update the exception testing code to test more exceptions; however,
note that the inst_fault exception is some what more difficult to
test here, since the PC gets updated before the fault is detected.
Thus returning to the EPC isn't viable.

With a quick hack it's easy to verify that instruction faults do
happen, but it's not so easy to recover gracefully from them. It
could be done, but I'm lazy, for now.

Also fix a sneaky bug in the trap entrance code. Use a bgez instead
of a bgtz since a value in MCAUSE of 0 corresponds to an instruction
alignment fault. Without this we hit an unknown interrupt failure.

Signed-off-by: Alex Waterman <imnotlistening@gmail.com>