r5sim: Add some optimization to the PMP check
Precompute as much of the PMP as possible when a PMP CSR is
updated. This allows us to spend less time in the check every
time an instruction is loaded or a register is loaded/stored.

It doesn't make a huge difference - the PMP is still a net loss
for perf, but it's better than nothing.

Signed-off-by: Alex Waterman <imnotlistening@gmail.com>